first

+++ app/src/main/java/com/bassem/shoopinguser/models/ItemsClass.kt
@@ -1,10 +1,10 @@
 package com.bassem.shoopinguser.models
 
 data class ItemsClass(
-    val title:String?=null,
-    val photo:String?=null,
-    val price:String?=null,
-    var favorite:Boolean=false,
-    val amount:Long?=null,
-    val id :String?=null
+    val title: String? = null,
+    val photo: String? = null,
+    val price: String? = null,
+    var favorite: Boolean = false,
+    val amount: Int? = null,
+    val id: String? = null
 )
